Move MySQL snapshot repository into `saga` subpackage.
