+++ src/axmysql/pkg.go
@@ -1,2 +1,3 @@
-// Package axmysql provides a MySQL-based persistence layer.
+// Package axmysql provides MySQL-backed implementations of various interfaces
+// consumed by Ax.
 package axmysql
