+++ src/axmysql/saga/pkg.go
@@ -1,0 +1,3 @@
+// Package saga provides MySQL-specific implementations of the interfaces in
+// Ax's top-level "saga" package.
+package saga
