+++ src/axmysql/saga/snapshot.go
@@ -1,4 +1,4 @@
-package axmysql
+package saga
 
 import (
 	"context"
